move all services/integration/config stuff to provider package
there were confusing interdependencies between these subsystems
