--- hibernate-reactive-core/src/main/java/org/hibernate/reactive/boot/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Integration with Hibernate ORM, via the service architecture.
- */
-package org.hibernate.reactive.boot;
