+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/provider/package-info.java
@@ -1,0 +1,12 @@
+/**
+ * Integration with Hibernate ORM, via the service architecture, along
+ * with APIs for starting and configuring Hibernate Reactive.
+ *
+ * <ul>
+ * <li>{@link org.hibernate.reactive.provider.ReactivePersistenceProvider}
+ *     is the JPA persistence provider for Hibernate Reactive.
+ * <li>{@link org.hibernate.reactive.provider.ReactiveServiceRegistryBuilder},
+ *     may be used to programmatically configure Hibernate Reactive.
+ * </ul>
+ */
+package org.hibernate.reactive.provider;
