+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/provider/service/ReactiveMarkerServiceSingleton.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright: Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.reactive.boot.impl.marker;
+package org.hibernate.reactive.provider.service;
 
 public final class ReactiveMarkerServiceSingleton implements ReactiveMarkerService {
