+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/provider/service/ReactiveSessionFactoryBuilder.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright: Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.reactive.boot.impl;
+package org.hibernate.reactive.provider.service;
 
 import org.hibernate.SessionFactory;
 import org.hibernate.boot.SessionFactoryBuilder;
