+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/session/package-info.java
@@ -1,4 +1,5 @@
 /**
- * Defines internal contracts between subsystems of Hibernate Reactive.
+ * Defines internal contracts between subsystems of Hibernate Reactive
+ * and the {@link org.hibernate.reactive.session.ReactiveSession}.
  */
 package org.hibernate.reactive.session;
