+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/vertx/impl/ProvidedVertxInstance.java
@@ -7,7 +7,6 @@ package org.hibernate.reactive.vertx.impl;
 
 import io.vertx.core.Vertx;
 
-import org.hibernate.reactive.boot.ReactiveServiceRegistryBuilder;
 import org.hibernate.reactive.vertx.VertxInstance;
 
 import java.util.Objects;
@@ -16,8 +15,8 @@ import java.util.Objects;
  * An implementation of {@link VertxInstance} which allows the client
  * to provide an instance of {@link Vertx} whose lifecyle is managed
  * externally to Hibernate Reactive. The {@code ProvidedVertxInstance}
- * must be registered with explicitly Hibernate by calling
- * {@link ReactiveServiceRegistryBuilder#addService}.
+ * must be registered explicitly with Hibernate by calling
+ * {@link org.hibernate.reactive.provider.ReactiveServiceRegistryBuilder#addService}.
  * Hibernate will destroy the {@code Vertx} instance on shutdown.
  *
  * <pre>
