+++ hibernate-reactive-core/src/main/resources/META-INF/services/org.hibernate.boot.model.TypeContributor
@@ -1,1 +1,1 @@
-org.hibernate.reactive.boot.impl.ReactiveTypeContributor
+org.hibernate.reactive.provider.impl.ReactiveTypeContributor
