+++ hibernate-reactive-core/src/test/java/org/hibernate/reactive/BaseMutinyTest.java
@@ -8,7 +8,7 @@ package org.hibernate.reactive;
 import org.hibernate.boot.registry.StandardServiceRegistry;
 import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.cfg.Configuration;
-import org.hibernate.reactive.boot.ReactiveServiceRegistryBuilder;
+import org.hibernate.reactive.provider.ReactiveServiceRegistryBuilder;
 import org.hibernate.reactive.containers.DatabaseConfiguration;
 import org.hibernate.reactive.mutiny.Mutiny;
 import org.hibernate.reactive.pool.ReactiveConnection;
