+++ hibernate-reactive-core/src/test/java/org/hibernate/reactive/BaseReactiveTest.java
@@ -11,8 +11,8 @@ import org.hibernate.boot.registry.StandardServiceRegistry;
 import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
-import org.hibernate.reactive.boot.ReactiveServiceRegistryBuilder;
-import org.hibernate.reactive.boot.service.ReactiveGenerationTarget;
+import org.hibernate.reactive.provider.ReactiveServiceRegistryBuilder;
+import org.hibernate.reactive.provider.service.ReactiveGenerationTarget;
 import org.hibernate.reactive.containers.DatabaseConfiguration;
 import org.hibernate.reactive.containers.DatabaseConfiguration.DBType;
 import org.hibernate.reactive.pool.ReactiveConnection;
